Renamed SomaModel.nb to CubicSomaModel.nb
This file used to be both Quadratic and Cubic Soma Models.  Before this
git repo was created, the Quadratic portion of this file was copied and
pasted into a separate file and then updated/upgraded from there.  That
file is now called QuadraticSomaModel.nb.  As such, I thought it would
be cleaner to rename and eventually repurpose this file to just
modelling the cubic soma model.