Low Pass Filter Manipulator added to SynapseModel.nb
A comprehensive lpf manipulator was added to SynapseModel.nb to show the
derivation process and math behind the basics of the the LPF
calculations and equations.  This manipulator can also be used to better
understand the effects of the time constant on the exponential curves.

The other two files have had some minor updates and this commit is also
to keep a track of those minor updates, in order to capture the latest
state of the mathematica scripts.