Updated TimeConstantCalculator to fit using exponential
Previously, the TimeConstantCalculator was fitting by taking the log of
the exponential data curve subtracted by the offset, and then fitting a
linear fit.  This new version takes the data as is, and fits the offset,
the Isat, and the tau value.  The Isat value that it gets is obviously
wrong, but the tau value is almost exactly 10ms when the circuit is
setup as such.  As a time constant calculator, this is definitely the
more accurate method to compute the value of tau, the time constant.