Added MismatchedSynapseFeedbackCkt.nb
This notebook is something that I whipped up really quickly with Sam to
help model Hakan's analysis of the feedback circuit.  It essentially
creates a manipulator that allows the user to vary the respective time
contants of the input synapse and the feedback synapse to see how any
mismatch between them affects the output.