Added a new version of the TimeConstantCalculator
This version allows the user to export multiple I(vmeas_im) curves in
the same waveform and it will parse all of the data at once. The
manipulator can now handle multiple taus now and update dynamically
which plot it is showing.

There are some minor bugs with this notebook:
1) It requires the user to evaluate the cell twice to truly reset the
state.
2) During the first run-through it needs to be evaluated twice as
well...otherwise it runs an infinite loop of errors.