Added SomaPTauCalc
SomaPTauCalc is a notebook that accepts an input file, gotten from the
output of SomaTauMeasPreparser.pl,  which then plots the simulated Tau
calibration results on the axes 1/f vs h(iin)/Ilk.  The slope of this
plot is fit to get a value for ptau.

Also included with this commmit is the perl script that takes data from
*.mres format (output from HSPICE's measurment saving feature) and
parses it for the important information.  It then outputs this
information into a simple txt file that the Mathematica script above
parses in order to calculate the ptau information.