LPFTimeConstCalc was used to model linear slope curves
In the LPF simulation, we used a linear fit of the log of the
exponential to get a slope that correlated with tau.  This was
originally done with just the data alone, but in looking back at the
simulation, it was noted that I_lk change as the experiment was
occuring.  What this implies is that tau is changing as time progresses.
As such, a theoretical model was added to LPFTimeConstCalc to model how
changes in tau over time affect the slope of the linearly fitted
function.

Also included in this commit are minor changes, including
SomaTauMeasPreparser.pl now outputting to the terminal it's output of
data as well.  This allows the user to quickly see what data was
compiled.  The latest version of SomaPTauCalc.nb is also included with
this commit.