Creation of SomaPQuaCalc.nb
This document was originally adapted from SomaPTauCalc.nb, but
eventually used virtually none of the elements of SomaPTauCalc.nb b/c it
never automatically grabbed data from a file.  Instead, as a user, I
manually added values to the document and then used Mathematica to plot
the bifurcation line for a simulated neuron with identically-sized
transistors.  This is used to verify that the calibration process works
in a controlled, repeatable manner.