SomaPQuaCalc includes data analysis from looking at longer time window
The SomaPQuaCalc script includes data analysis from trying to simulate
the bifurcation process on a Soma with no Series transistor (M62), no
Van circuit, including the biases from the circuit side (as opposed to
the bias side, except for Iin), and using a simulation window of
100seconds to better assess when a neuron bifurcates.

With this data, the notebook now also includes a pseudo-validation step
that checks what value of i_in the neuron bifurcates when a pqua and b
(offset) value are also used to calculate Iin for a given Ilk.  Now the
neuron bifurcates for i_in values ranging between .5071 to .5395.