Added BiasParamTranslator.nb
This new script allows me to quickly convert between dimensionless and
dimensional forms with known calibration params. Currently it only
implements Soma Calib conversions.  Also, it assumes the simplified
dual-bias form of the circuit, as opposed to stacked bias form

This commit also includes the latest status and updates to the other
files included in this commit.  These files include updates where a lot
of modelling and analysis was performed on the Quadratic Soma and how
h[vin] changes and operates.