New Mathematica scripts
DAC Calculator gives a mathematica based way to convert from DAC bits to
actual current values using the values set in the DAC bias file

FiringRateTheory covers the theory behind doing series approximations
for calculating the firing rate

SomaCalibTheory is a notebook that addresses the effects of pqua, ptau,
and pilkr on the firing rate and calibration process.  Included in this
is an evaluation of why using the median neuron doesn't make sense.
This uses approximations for the distribution of the neurons based on
pqua and ptau values.