FiringRateTheory now includes the theory behind T^2 based calibration
This notebook now includes the derivation and theory behind using the
period, instead of the firing rate, for calibrating in the new unified
calibration process.  This new method (using the period) is being
investigated because we feel that it might help with the tau sensitivity
in the firing rate method.  The tau term is sensitive in the old method
because a lot of the firing rates are near 0, at bifurcation, meaning
that we get very little "resolution" at that point.  Taking the inverse
of the firing rate means that more of our data will be in a regime where
the resolution of the error will be much higher.  This will hopefully
resolve the errors and variance we were seeing in the ptau calibration.