FiringRateTheory now has 3rd order approximations included
The FiringRateTheory file now includes math, theory, and plots behind
using a 3rd order approximation of f^2.  It also has an approximate
error calculator that gives what the expected error is for a given iin
value.

Also included in this commit is an update to SomaCalibTheory (as well as
FiringRateTheory) where all contour plots now have labels on the axes.
Instead of using AxesLabel as the command it is necessary to use
FrameLabel for contour plots.

Lastly, included in this commit is a snapshot of QuadraticSomaModel as
of this commit.