Initial Sensitivity Analysis math added
Also included in this commit is a fix/change that avoids renaming
errors when calculating the approximate error of various order series
approximations.  This problem with the old notebook was causing me to
incorrectly see that the 3rd Order Series Approximation of f^2 was more
erroneous than the 2nd Order Series Approximation.  This is now correct
and consistent.