Consolidated and Updated/Upgraded version of RefractoryPeriodModel
Also included in this commit are updates/upgrades made to the
FiringRateTheory.

RefractoryPeriodModel now shows the derivation and explanation for how
to solve for V_r in terms of other voltages and currents in the circuit.
It also has a nice standardized way of representing many of the Soma
equations that I should/will eventually transfer to other notebooks or
to a centralized notebook that all notebooks can access.