Mismatch analysis for AD-TDC implemented
This version of ADC Mismatch now has analysis for the mismatch as it
applies to an Analog to Digital Time-to-Digital Converter.  More
specifically, the analysis is performed on the ring-oscillator to see
how much each inverter varies relative to other inverters in the chain.
We wanted to ensure that the variance is between 0.5LSB and 1.5LSB.